requirement.txt
astroid==2.11.5
autopep8==1.6.0
CacheControl==0.12.11
cachelib==0.7.0
cachy==0.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cleo==0.8.1
click==8.1.3
clikit==0.6.2
colorclass==2.2.2
crashtest==0.3.1
cryptography==37.0.2
dill==0.3.5.1
distlib==0.3.4
docopt==0.6.2
filelock==3.7.0
Flask==2.1.2
Flask-Login==0.6.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
isort==5.10.1
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
keyring==23.5.1
lazy-object-proxy==1.7.1
lockfile==0.12.2
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.3
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pip-upgrader==1.4.15
pkginfo==1.8.2
platformdirs==2.5.2
poetry==1.1.13
poetry-core==1.0.8
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.21
pylev==1.4.0
pylint==2.13.9
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.27.1
requests-toolbelt==0.9.1
SecretStorage==3.3.2
shellingham==1.4.0
six==1.16.0
SQLAlchemy==1.4.39
terminaltables==3.1.10
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
webencodings==0.5.1
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
